removed middleware
