+++ src/App.js
@@ -1,6 +1,6 @@
 import React, { useEffect } from "react";
 import LandingPage from "./pages/LandingPage";
-import { Routes, Route, HashRouter } from "react-router-dom";
+import { Routes, Route, HashRouter, useLocation } from "react-router-dom";
 import Contact from "./pages/Contact";
 import About from "./pages/About";
 import GlobheApp from "./pages/GlobheApp";
@@ -18,6 +18,14 @@ const App = () => {
   //     gtmId: "G-M3XXNGRW2H",
   //   });
   // }, []);
+  const location = useLocation();
+
+  useEffect(() => {
+    window.gtag("event", "page_view", {
+      page_path: location.pathname + location.search,
+      page_location: window.location.href,
+    });
+  }, [location]);
 
   return (
     <HashRouter>
@@ -44,11 +52,16 @@ const App = () => {
           title="My resume"
           element={<Resume />}
         /> */}
-        <Route exact path="/" element={<LandingPage />} />
-        <Route path="/about" element={<About />} />
-        <Route path="/contact" element={<Contact />} />
-        <Route path="/globhe-app" element={<GlobheApp />} />
-        <Route path="/my-resume" element={<Resume />} />
+        <Route
+          exact
+          path="/"
+          title="Rémy Ghalayini"
+          element={<LandingPage />}
+        />
+        <Route path="/about" title="Contact me" element={<About />} />
+        <Route path="/contact" title="Contact me" element={<Contact />} />
+        <Route path="/globhe-app" title="GLOBHE app" element={<GlobheApp />} />
+        <Route path="/my-resume" title="My resume" element={<Resume />} />
       </Routes>
     </HashRouter>
   );
